requirements.txt
mcp==1.27.0
anthropic==0.28.0
pydantic>=2.11.0
pyyaml==6.0.1
requests==2.31.0
python-dotenv==1.0.0